Implement generate_udf_application_steps and move generators to single file
